refactor: move mate::Locker to gin
