+++ filenames.gni
@@ -433,8 +433,6 @@ filenames = {
     "shell/common/api/event_emitter_caller_deprecated.cc",
     "shell/common/api/event_emitter_caller_deprecated.h",
     "shell/common/api/features.cc",
-    "shell/common/api/locker.cc",
-    "shell/common/api/locker.h",
     "shell/common/asar/archive.cc",
     "shell/common/asar/archive.h",
     "shell/common/asar/asar_util.cc",
@@ -500,6 +498,8 @@ filenames = {
     "shell/common/gin_helper/event_emitter.h",
     "shell/common/gin_helper/function_template.cc",
     "shell/common/gin_helper/function_template.h",
+    "shell/common/gin_helper/locker.cc",
+    "shell/common/gin_helper/locker.h",
     "shell/common/gin_helper/object_template_builder.cc",
     "shell/common/gin_helper/object_template_builder.h",
     "shell/common/gin_helper/trackable_object.cc",
