+++ shell/common/api/electron_bindings.cc
@@ -22,9 +22,9 @@
 #include "services/resource_coordinator/public/cpp/memory_instrumentation/global_memory_dump.h"
 #include "services/resource_coordinator/public/cpp/memory_instrumentation/memory_instrumentation.h"
 #include "shell/browser/browser.h"
-#include "shell/common/api/locker.h"
 #include "shell/common/application_info.h"
 #include "shell/common/gin_helper/dictionary.h"
+#include "shell/common/gin_helper/locker.h"
 #include "shell/common/heap_snapshot.h"
 #include "shell/common/native_mate_converters/file_path_converter.h"
 #include "shell/common/native_mate_converters/string16_converter.h"
@@ -131,7 +131,7 @@ void ElectronBindings::OnCallNextTick(uv_async_t* handle) {
            self->pending_next_ticks_.begin();
        it != self->pending_next_ticks_.end(); ++it) {
     node::Environment* env = *it;
-    mate::Locker locker(env->isolate());
+    gin_helper::Locker locker(env->isolate());
     v8::Context::Scope context_scope(env->context());
     node::InternalCallbackScope scope(
         env, v8::Local<v8::Object>(), {0, 0},
@@ -236,7 +236,7 @@ v8::Local<v8::Promise> ElectronBindings::GetProcessMemoryInfo(
   util::Promise<mate::Dictionary> promise(isolate);
   v8::Local<v8::Promise> handle = promise.GetHandle();
 
-  if (mate::Locker::IsBrowserProcess() && !Browser::Get()->is_ready()) {
+  if (gin_helper::Locker::IsBrowserProcess() && !Browser::Get()->is_ready()) {
     promise.RejectWithErrorMessage(
         "Memory Info is available only after app ready");
     return handle;
@@ -271,7 +271,7 @@ void ElectronBindings::DidReceiveMemoryDump(
     bool success,
     std::unique_ptr<memory_instrumentation::GlobalMemoryDump> global_dump) {
   v8::Isolate* isolate = promise.isolate();
-  mate::Locker locker(isolate);
+  gin_helper::Locker locker(isolate);
   v8::HandleScope handle_scope(isolate);
   v8::MicrotasksScope script_scope(isolate,
                                    v8::MicrotasksScope::kRunMicrotasks);
